requirements.txt
crewai==0.28.8
crewai_tools==0.1.6
langchain==0.1.11
mistralai==1.5.0
python-dotenv==1.0.0